+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## UNRELEASED
 
+- Fixed bug where ActiveRecordConsumer was not using `unscoped` to update
+  via primary key and causing duplicate record errors.
+
 # [1.1.0-beta1] - 2019-09-10
 - Added BatchConsumer.